chore: add wrapper router for sites v1
